% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cpi_combined.R
\docType{data}
\name{df_cpi_combined}
\alias{df_cpi_combined}
\title{Consumer Price Index (CPI) data from CBS}
\format{
\subsection{\code{df_cpi_combined}}{

A data frame with 11 rows and 8 columns:
\describe{
\item{Year from}{Year starting from}
\item{Year to}{Year ending}
\item{Percentage}{Percentage change}
\item{Factor}{Factor for multiplication}
\item{Friction period in days}{Calculation of the friction period in days}
\item{Friction period in weeks}{Calculation of the friction period in weeks}
\item{Friction period days average over 5 years}{Calculation of the 5 year average friction period in days}
\item{Friction period weeks average over 5 years}{Calculation of the 5 year average friction period in weeks}
}
}
}
\source{
\url{https://www.cbs.nl/nl-nl/cijfers}
}
\usage{
df_cpi_combined
}
\description{
A subset data frame of Consumentenprijzen; prijsindex 2015=100 from CBS. Identifier: 83131NED
}
\keyword{datasets}
