% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fp.R
\docType{data}
\name{df_fp}
\alias{df_fp}
\title{Job vacancy data from CBS}
\format{
\subsection{\code{df_fp}}{

A data frame with 27 rows and 7 columns:
\describe{
\item{Year}{Year}
\item{Filled vacancies}{Vervulde vacatures}
\item{Open vacancies}{Openstaande vacatures}
\item{Friction period in days}{Calculation of the friction period in days}
\item{Friction period in weeks}{Calculation of the friction period in weeks}
\item{Friction period days average over 5 years}{Calculation of the 5 year average friction period in days}
\item{Friction period weeks average over 5 years}{Calculation of the 5 year average friction period in weeks}
}
}
}
\source{
\url{https://www.cbs.nl/nl-nl/cijfers}
}
\usage{
df_fp
}
\description{
A subset data frame of job vacancies; SBI 2008; by economic activity and company size from CBS. Identifier: 80472NED
}
\keyword{datasets}
