% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ppp.R
\docType{data}
\name{df_ppp}
\alias{df_ppp}
\title{Purchasing Power Parity (PPP) data from OECD}
\format{
\subsection{\code{df_ppp}}{

A data frame with 64 rows and 2 columns:
\describe{
\item{Year}{Year of PPP}
\item{PPP}{Purchase Power Parity}
}
}
}
\source{
\url{https://sdmx.oecd.org/public}
}
\usage{
df_ppp
}
\description{
A subset data frame of the OECD data set containing the Purchasing Power Parity (PPP) data.
}
\keyword{datasets}
