% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_df_ref_prices.R
\docType{data}
\name{df_ref_prices}
\alias{df_ref_prices}
\title{Complete data of reference values from the costing manual}
\format{
\subsection{\code{df_ref_prices}}{

A data frame with 160 rows and 7 columns:
\describe{
\item{Domain}{Care domain, for example medical services}
\item{Category}{Service category such as nursing, intensive care or day treatment}
\item{Unit}{Description of the care unit or service}
\item{short_var}{Short variable name used in the package}
\item{2022}{Reference price for the year 2022}
\item{2023}{Reference price for the year 2023}
\item{2024}{Reference price for the year 2024}
}
}
}
\source{
\url{https://sdmx.oecd.org/public}
}
\usage{
df_ref_prices
}
\description{
Complete data of reference values from the costing manual
}
\keyword{datasets}
