% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rp_medical.R
\docType{data}
\name{df_rp_medical}
\alias{df_rp_medical}
\title{Medical unit cost data from the Costing manual: Methods and Reference Prices for Economic Evaluations in Healthcare}
\format{
\subsection{\code{df_rp_medical}}{

A data frame with 116 rows and 4 columns:
\describe{
\item{Category}{Category of the medical unit cost}
\item{Unit}{Medical unit cost}
\item{2022}{Year 2022 medical unit cost}
\item{2023}{Year 2023 medical unit cost}
}
}
}
\source{
<Hakkaart - van Roijen, L., Peeters, S., Kanters, T., van Baal, P., Brouwer, W., Drost, R., Evers, S. M. A. A., van Exel, J., Reckers-Droog, V., Tannaoui, N.-E., Thielen, F., & Wijnen, B. (2024). Kostenhandleiding voor economische evaluaties in de gezondheidszorg: Methodologie en Referentieprijzen. (Herziene versie 2024 ed.).>
}
\usage{
df_rp_medical
}
\description{
A subset data frame of medical unit cost data; chapter 4 of the Dutch Costing Manual
}
\keyword{datasets}
