% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rp_prod.R
\docType{data}
\name{df_rp_prod}
\alias{df_rp_prod}
\title{Productivity & other unit cost data from the Costing manual: Methods and Reference Prices for Economic Evaluations in Healthcare}
\format{
\subsection{\code{df_rp_prod}}{

A data frame with 38 rows and 4 columns:
\describe{
\item{Category}{Category of the productivity & other unit cost}
\item{Unit}{Productivity & other unit cost}
\item{2022}{Year 2022 productivity & other unit cost}
\item{2023}{Year 2023 productivity & other unit cost}
}
}
}
\source{
<Hakkaart - van Roijen, L., Peeters, S., Kanters, T., van Baal, P., Brouwer, W., Drost, R., Evers, S. M. A. A., van Exel, J., Reckers-Droog, V., Tannaoui, N.-E., Thielen, F., & Wijnen, B. (2024). Kostenhandleiding voor economische evaluaties in de gezondheidszorg: Methodologie en Referentieprijzen. (Herziene versie 2024 ed.).>
}
\usage{
df_rp_prod
}
\description{
A subset data frame of productivity & other unit cost data; chapter 6 of the Dutch Costing Manual
}
\keyword{datasets}
