% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount_value.R
\name{discount_value}
\alias{discount_value}
\title{A function to calculate the discounted value of future costs or effects}
\usage{
discount_value(
  value,
  year,
  discount_rate = c("costs", "effects"),
  discount_year_one = FALSE
)
}
\arguments{
\item{value}{The value to be discounted.}

\item{discount_rate}{Either "costs", "effects", or a numeric rate. Default is "costs".}

\item{discount_year_one}{Logical value indicating whether to discount the first year as well. Default is FALSE.}

\item{time}{The time at which the future value occurs.}
}
\value{
A numeric value of the discounted value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calculates the discounted value of a future costs or effects based on the in paragraph 2.6.1.2 of the Dutch EE guideline mentioned discount rate and time period
}
\examples{
# Example usage of the discount_value function
# Calculate the discounted value of 100 after 3 years, the first year is not discounted
discount_value(value = 100, discount_rate = 0.03, year = 3)

}
\keyword{Costs,}
\keyword{Discounting,}
\keyword{Effects}
