% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_med_prices.R
\name{nl_med_prices}
\alias{nl_med_prices}
\title{A function to download the Medical Reference prices of the Dutch Costing Manual for one or multiple years}
\usage{
nl_med_prices(year = "all", category = "all", currency = "EUR", unit = "all")
}
\arguments{
\item{year}{The year of which the reference price should be downloaded, multiple years are possible, default is the whole dataset}

\item{category}{The category of prices that should be included (one or more categories), default is including all categories}

\item{currency}{The currency of the output of the prices. A decision can be made between EUR and INT$, the default is EUR.}

\item{unit}{The reference price that should be included (one or multiple reference prices),  default is including the whole dataframe}
}
\value{
A dataframe with the Medical Reference prices of the Dutch Costing Manual for the specified years.
}
\description{
This function downloads the Medical Reference prices of the Dutch Costing Manual for one or multiple years. The prices are available in Euro (EUR) or International Dollar (INT$).
}
\examples{
Example usage of the nl_med_prices function
nl_med_prices(year = 2019, category = "Nursing", unit = "Nursing day excluding personnel costs, hospital")
}
\keyword{Costing}
\keyword{Dutch}
\keyword{Generic,}
\keyword{Manual,}
\keyword{Medical}
\keyword{Prices}
\keyword{Prices,}
\keyword{Reference}
