% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_ppp.R
\name{nl_ppp}
\alias{nl_ppp}
\title{A function to obtain the Dutch PPP factor values in International Dollar (INT$)}
\usage{
nl_ppp(year = "all")
}
\arguments{
\item{year}{The year of which the PPP factor should be downloaded, multiple years are possible, default is the whole dataset.}
}
\value{
A dataframe or value with the PPP factor values for the specified years.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function downloads the Purchasing Power Parity (PPP) factor values for the Netherlands from the OECD website per year in International Dollar (Int$).
}
\examples{
# Example usage of the nl_ppp function
nl_ppp(year = 2019)
}
\keyword{PPP}
\keyword{Costing}
\keyword{Generic}
\keyword{Manual}
\keyword{Parity}
\keyword{Power}
\keyword{Purchasing}
