% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_price_index.R
\name{nl_price_index}
\alias{nl_price_index}
\title{A function to calculate the Consumer Price Index (CPI) for a given year range.}
\usage{
nl_price_index(start_year = 2013, end_year = 2023, output = "table")
}
\arguments{
\item{start_year}{start year for CPI output table}

\item{end_year}{End year for CPI output table}

\item{output}{Which output we would like to see. "factor": is the factor from start to end year, "table" is the table of all CPIs from start to end year}
}
\value{
dataframe with CPI data (CPI, percentage and multiplication factor) from start year to end year
}
\description{
This function provides the Consumer Price Index (CPI) for a given year range both in a factor or dataframe
}
\examples{
example usage of the nl_price_index function
nl_price_index(2015, 2019, output = "table")

}
