% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_price.R
\name{pretty_price}
\alias{pretty_price}
\title{A function to write pretty prices in \code{bookdown} reports}
\usage{
pretty_price(x, digi = 2, currency = "EUR", ...)
}
\arguments{
\item{x}{A number to be printed}

\item{digi}{Number of digits}

\item{currency}{The name of the currency}

\item{...}{Extra arguments for \code{formatC()}}
}
\value{
A pretty price with the currency
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function writes pretty prices in \code{bookdown} reports. The function uses the \code{formatC()} function to format the number and adds the currency to the end of the number.
}
\examples{
# Example usage of the pretty_price function
pretty_price(1000, currency = "EUR")
}
