% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friction_period.R
\name{friction_period}
\alias{friction_period}
\title{A function to download the friction period over one or multiple years}
\usage{
friction_period(year = "all", period = "all", type = "5_year_avg")
}
\arguments{
\item{year}{The year of which the friction period should be downloaded, multiple years are possible. The default is the whole data frame}

\item{period}{The friction period that should be included (days/weeks),  default is including the whole dataframe}

\item{type}{If period is chosen, a decision can be made between the 5 year average and 1 year friction period in days/weeks. The default is the 5 year average}
}
\value{
A data frame with friction periods for all years or a selection of years and variables.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Example usage of the depreciation_interest function
friction_period(year = 2019, period = "weeks", type = "5_year_avg")
}
\keyword{CBS}
\keyword{Generic}
